third_party/WebKit/LayoutTests/web-animations-api/reject-hyphen-expected.txt

CONSOLE WARNING: Hyphenated property names in Web Animations keyframes are invalid and therefore ignored. Please use camelCase instead.
This is a testharness.js-based test.
PASS Hyphenated propery names are ignored. 
Harness: the test ran to completion.
